requirements.txt
spotipy==2.24.0
requests==2.32.3
pillow==10.3.0
Flask==3.0.3